% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_functions.R
\name{calc_dominant_party_family}
\alias{calc_dominant_party_family}
\title{Calculate Dominante Party Family}
\usage{
calc_dominant_party_family(
  data,
  party_A_vote_share,
  party_B_vote_share,
  party_C_vote_share,
  party_D_vote_share,
  party_E_vote_share,
  party_F_vote_share,
  party_A_family,
  party_B_family,
  party_C_family,
  party_D_family,
  party_E_family,
  party_F_family
)
}
\arguments{
\item{data}{a data frame}
}
\description{
Calculate Dominante Party Family
}
