% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_left_right_scale}
\alias{clean_left_right_scale}
\title{Clean Left Right Scale}
\usage{
clean_left_right_scale(data, var_name)
}
\arguments{
\item{data}{a data frame}

\item{var_name}{name of column with left right scale}
}
\description{
This function should work for self identified left right, expert identified left right,
and respondent classification of all parties across all modules.
}
