% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cses_subset.R
\docType{data}
\name{cses_subset}
\alias{cses_subset}
\title{A Subset of data from the CSES}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 317551 rows and 88 columns.}
\usage{
cses_subset
}
\description{
A subset of data from Modules 1,2,3,4 and the advanced release of Module 5. Columns selected were common between multiple
}
\keyword{datasets}
