% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_election_type}
\alias{decode_election_type}
\title{decode election type encoding}
\usage{
decode_election_type(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with election type}
}
\value{
a tbl
}
\description{
decode election type encoding
}
