% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_employment_status}
\alias{decode_employment_status}
\title{decode employment status}
\usage{
decode_employment_status(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with employment status}
}
\description{
decode employment status
}
