% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_household_income}
\alias{decode_household_income}
\title{decode Household Income}
\usage{
decode_household_income(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Household Income}
}
\description{
decode Household Income
}
