% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_marital_status}
\alias{decode_marital_status}
\title{decode marital status}
\usage{
decode_marital_status(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with gender encoding}
}
\description{
decode marital status
}
