% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_party_family}
\alias{decode_party_family}
\title{decode Party Ideological Family}
\usage{
decode_party_family(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with party_family}
}
\description{
decode Party Ideological Family
}
