% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_residence_type}
\alias{decode_residence_type}
\title{decode Residence Type}
\usage{
decode_residence_type(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Residence Type}
}
\description{
decode Residence Type
}
