% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_satisfaction}
\alias{decode_satisfaction}
\title{decode Satisfaction with democracy}
\usage{
decode_satisfaction(data, var_name, module_var)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Satisfaction with democracy}
}
\description{
decode Satisfaction with democracy
}
