% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{decode_survey_timing}
\alias{decode_survey_timing}
\title{Decode election timing encoding}
\usage{
decode_survey_timing(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with election type}
}
\description{
Decode election timing encoding
}
