% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{reduce_employment_status}
\alias{reduce_employment_status}
\title{Reduce the factor levels of employment status}
\usage{
reduce_employment_status(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with employment status}
}
\description{
Reduce the factor levels of employment status
}
