% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_education}
\alias{translate_education}
\title{Translate education status}
\usage{
translate_education(data, var_name, module_var)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with gender encoding}

\item{module_var}{name of column with module information}
}
\description{
I've made some opinionated recoding of module 1,2,&3's coding to bring them in line with 4/5
}
