% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_election_type}
\alias{translate_election_type}
\title{Translate election type encoding}
\usage{
translate_election_type(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with election type}
}
\value{
a tbl
}
\description{
Translate election type encoding
}
