% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_employment_status}
\alias{translate_employment_status}
\title{Translate employment status}
\usage{
translate_employment_status(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with employment status}
}
\description{
Translate employment status
}
