% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_household_income}
\alias{translate_household_income}
\title{Translate Household Income}
\usage{
translate_household_income(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Household Income}
}
\description{
Translate Household Income
}
