% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_marital_status}
\alias{translate_marital_status}
\title{Translate marital status}
\usage{
translate_marital_status(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with gender encoding}
}
\description{
Translate marital status
}
