% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_party_family}
\alias{translate_party_family}
\title{Translate Party Ideological Family}
\usage{
translate_party_family(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with party_family}
}
\description{
Translate Party Ideological Family
}
