% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_power_matters}
\alias{translate_power_matters}
\title{Translate Belief that Who is in power matters}
\usage{
translate_power_matters(data, var_name)
}
\arguments{
\item{data}{a data frame with CSES columns}

\item{var_name}{name of column containting raw response to question of if who is in power matters}
}
\value{
a tbl
}
\description{
Translate Belief that Who is in power matters
}
