% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_residence_type}
\alias{translate_residence_type}
\title{Translate Residence Type}
\usage{
translate_residence_type(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Residence Type}
}
\description{
Translate Residence Type
}
