% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_responder_gender}
\alias{translate_responder_gender}
\title{Translate Respondent Gender}
\usage{
translate_responder_gender(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with gender encoding}
}
\value{
a tbl
}
\description{
Translate Respondent Gender
}
