% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{translate_satisfaction}
\alias{translate_satisfaction}
\title{Translate Satisfaction with democracy}
\usage{
translate_satisfaction(data, var_name, module_var)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with Satisfaction with democracy}
}
\description{
Translate Satisfaction with democracy
}
