% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_functions.R
\name{calc_labor_force_status}
\alias{calc_labor_force_status}
\title{Calculate Labor Force Status}
\usage{
calc_labor_force_status(data, var_name, new_name)
}
\arguments{
\item{data}{A data frame}

\item{var_name}{Name of column with employment status}

\item{new_name}{Name of new column}
}
\value{
a tbl with a new column for labor force
}
\description{
Calculate Labor Force Status
}
