% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_household_size}
\alias{clean_household_size}
\title{clean household_size}
\usage{
clean_household_size(data, var_name)
}
\arguments{
\item{data}{a data frame from CSES}

\item{var_name}{the name of the column with household_size}
}
\description{
clean household_size
}
